feat: remove backend pagination for members & invitations
