+++ javascript/apps/taiga/src/app/modules/workspace/feature-detail/components/workspace-detail-people-members/workspace-detail-people-members.component.html
@@ -39,6 +39,7 @@ Copyright (c) 2023-present Kaleidos INC
                 class="user"
                 [@removeCell]="member.cancelled">
                 <tg-user-card
+                  [isSelf]="vm.currentUser.username === member.user.username"
                   [user]="member.user"
                   [navigateToUser]="true"></tg-user-card>
               </tg-ui-dtable-cell>
